requirements.txt
gradio>=4.0.0
faster-whisper
opencv-python-headless
pytubefix
numpy
python-multipart
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
